+++ src/NeuralNet/Initializers/Xavier2.php
@@ -32,6 +32,6 @@ class Xavier2 implements Initializer
     public function initialize(int $fanIn, int $fanOut) : Matrix
     {
         return Matrix::uniform($fanOut, $fanIn)
-            ->multiply((6. / ($fanOut + $fanIn)) ** 0.25);
+            ->multiply((6.0 / ($fanOut + $fanIn)) ** 0.25);
     }
 }