+++ tests/FunctionsTest.php
@@ -9,6 +9,7 @@ use function Rubix\ML\argmin;
 use function Rubix\ML\argmax;
 use function Rubix\ML\logsumexp;
 use function Rubix\ML\array_transpose;
+use function Rubix\ML\comb;
 
 /**
  * @group Functions
@@ -90,5 +91,32 @@ class FunctionsTest extends TestCase
                 [0],
             ],
         ];
+    }
+
+    /**
+     * @test
+     * @dataProvider combProvider
+     *
+     * @param int $n
+     * @param int $k
+     * @param int $expected
+     */
+    public function comb(int $n, int $k, int $expected) : void
+    {
+        $this->assertEquals($expected, comb($n, $k));
+    }
+
+    /**
+     * @return \Generator<array>
+     */
+    public function combProvider() : Generator
+    {
+        yield [1, 1, 1];
+
+        yield [2, 1, 2];
+
+        yield [8, 3, 56];
+
+        yield [10, 6, 210];
     }
 }