+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - Missing Data Imputer now passes through other data types
     - Changed order of Missing Data Imputer params
     - Renamed high-level resource type to image
+    - Added comb (n choose k) as a global function
 
 - 0.0.18-beta
     - Now requires PHP 7.2 and above